% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRefFreeEwasModel.R
\name{BootRefFreeEwasModel}
\alias{BootRefFreeEwasModel}
\title{Bootstrap for Reference-Free EWAS Model}
\usage{
BootRefFreeEwasModel(mod, nboot)
}
\arguments{
\item{mod}{model object of class RefFreeEwasModel (generated with smallOutput=FALSE).}

\item{nboot}{Number of bootstrap samples to generate.}
}
\value{
An array object of class \dQuote{BootRefFreeEwasModel}.
Bootstraps are generated for both Beta and Bstar.
}
\description{
Bootstrap procedure for reference-free method for conducting EWAS while
deconvoluting DNA methylation arising as mixtures of cell types.
}
\details{
Generates the bootstrap samples for the reference-free method
for conducting EWAS while deconvoluting DNA methylation arising
as mixtures of cell types.
}
\examples{

data(RefFreeEWAS)

if (interactive()) {
  tmpDesign <- cbind(1, rfEwasExampleCovariate)
  tmpBstar <- (rfEwasExampleBetaValues \%*\% tmpDesign \%*\% solve(t(tmpDesign)\%*\%tmpDesign))

  EstDimRMT(rfEwasExampleBetaValues-tmpBstar \%*\% t(tmpDesign))$dim
}

test <- RefFreeEwasModel(
  rfEwasExampleBetaValues,
  cbind(1,rfEwasExampleCovariate),
  4
)

testBoot <- BootRefFreeEwasModel(test,10)
summary(testBoot)
}
\seealso{
\code{RefFreeEwasModel}
}
\keyword{deconvolution,DNA}
\keyword{methylation,EWAS,surrogate}
\keyword{mixture,bootstrap}
\keyword{variable,cell}
