% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefFreeCellMix.R
\name{ImputeByMean}
\alias{ImputeByMean}
\title{Simple imputation method based on row-mean}
\usage{
ImputeByMean(Y)
}
\arguments{
\item{Y}{Matrix to impute.
'}
}
\value{
Matrix with missing values replaced by imputed values.
}
\description{
Simple method for imputing missing values by row-mean
}
