% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsBootOneRefFreeEwasModel.R
\name{PairsBootOneRefFreeEwasModel}
\alias{PairsBootOneRefFreeEwasModel}
\title{One Bootstrap Sample for Reference-Free EWAS Model, Accounting for Paired Data}
\usage{
PairsBootOneRefFreeEwasModel(mod, pairID)
}
\arguments{
\item{mod}{model object of class RefFreeEwasModel (generated with smallOutput=FALSE).}

\item{pairID}{Pair IDs (one unique value per pair).}
}
\value{
A matrix representing a bootstrap sample of an DNA methylation assay matrix.
}
\description{
Bootstrap generation procedure for reference-free method for conducting EWAS while deconvoluting DNA methylation arising as mixtures of cell types.
This version accounts for paired data (e.g. twin data)
}
\details{
Generates one bootstrapped data set for the reference-free method for conducting EWAS while deconvoluting DNA methylation arising as mixtures of cell types.  This version facilitates the estimation of robust standard errors to account for paired data (e.g. twin data) using a strategy similar to that employed by Generalized Estimating Equations (GEEs).  Specifically, in bootstrapping the errors, the pairs are sampled rather than individual arrays.  Typically not run by user.
}
\seealso{
\code{BootRefFreeEwasModel},\code{BootOneRefFreeEwasModel}
}
\keyword{deconvolution,DNA}
\keyword{methylation,EWAS,surrogate}
\keyword{mixture}
\keyword{variable,cell}
