% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsBootRefFreeEwasModel.R
\name{PairsBootRefFreeEwasModel}
\alias{PairsBootRefFreeEwasModel}
\title{Bootstrap for Reference-Free EWAS Model, Accounting for Paired Data}
\usage{
PairsBootRefFreeEwasModel(mod, nboot, pairID)
}
\arguments{
\item{mod}{model object of class RefFreeEwasModel (generated with smallOutput=FALSE).}

\item{nboot}{Number of bootstrap samples to generate.}

\item{pairID}{Pair IDs (one unique value per pair).}
}
\value{
An array object of class \dQuote{BootRefFreeEwasModel}. Bootstraps are generated for both Beta and Bstar.
}
\description{
Bootstrap procedure for reference-free method for conducting EWAS while deconvoluting DNA methylation arising as mixtures of cell types.
This version accounts for paired data (e.g. twin data)
}
\details{
Generates the bootstrap samples for the reference-free method for conducting EWAS while deconvoluting DNA methylation arising as mixtures of cell types.  This paired version facilitates the estimation of robust standard errors to account for paired data (e.g. twin data) using a strategy similar to that employed by Generalized Estimating Equations (GEEs).  Specifically, in bootstrapping the errors, the pairs are sampled rather than individual arrays.  An error will be generated unless each cluster has exactly two members (i.e. exactly two observations correspond to the same unique ID given in pairID).
}
\examples{

data(RefFreeEWAS)

if (interactive()) {
  tmpDesign <- cbind(1, rfEwasExampleCovariate)
  tmpBstar <- (rfEwasExampleBetaValues \%*\% tmpDesign \%*\% solve(t(tmpDesign)\%*\%tmpDesign))

  EstDimRMT(rfEwasExampleBetaValues-tmpBstar \%*\% t(tmpDesign))$dim
}

test <- RefFreeEwasModel(
  rfEwasExampleBetaValues,
  cbind(1,rfEwasExampleCovariate),
  4
)

testBoot <- BootRefFreeEwasModel(test,10)
summary(testBoot)
}
\seealso{
\code{RefFreeEwasModel},\code{BootRefFreeEwasModel}
}
\keyword{deconvolution,DNA}
\keyword{methylation,EWAS,surrogate}
\keyword{mixture,bootstrap}
\keyword{variable,cell}
