% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRefFreeEwasModel.R
\name{print.BootRefFreeEwasModel}
\alias{print.BootRefFreeEwasModel}
\title{print.BootRefFreeEwasModel}
\usage{
\method{print}{BootRefFreeEwasModel}(x, ...)
}
\arguments{
\item{x}{BootRefFreeEwasModel object to print}

\item{...}{Unused..}
}
\description{
Print method for objects of type BootRefFreeEwasModel
}
\details{
See \code{RefFreeEwasModel} for example.
}
