% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefFreeCellMix.R
\name{print.RefFreeCellMix}
\alias{print.RefFreeCellMix}
\title{print.RefFreeCellMix}
\usage{
\method{print}{RefFreeCellMix}(x, ...)
}
\arguments{
\item{x}{RefFreeCellMix object to print}

\item{...}{Unused..}
}
\description{
Print method for objects of type RefFreeCellMix
}
\details{
See \code{RefFreeCellMix} for example.
}
