% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefFreeEWAS.R
\name{print.RefFreeEwasModel}
\alias{print.RefFreeEwasModel}
\title{print.RefFreeEwasModel}
\usage{
\method{print}{RefFreeEwasModel}(x, ...)
}
\arguments{
\item{x}{RefFreeEwasModel object to print}

\item{...}{Unused..}
}
\description{
Print method for objects of type RefFreeEwasModel
}
\details{
See \code{RefFreeEwasModel} for example.
}
