\name{print.summaryBootRefFreeEwasModel}
\alias{print.summaryBootRefFreeEwasModel}
\title{print.summaryBootRefFreeEwasModel}
\description{
Print method for objects of type summaryBootRefFreeEwasModel
}
\usage{
\method{print}{summaryBootRefFreeEwasModel}(x,...)
}
\arguments{
  \item{x}{summaryBootRefFreeEwasModel object to print}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{RefFreeEwasModel}} for example.
}
