% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefFreeEWAS.R
\name{RefFreeEwasModel}
\alias{RefFreeEwasModel}
\title{Reference-Free EWAS Model}
\usage{
RefFreeEwasModel(Y, X, K, smallOutput = FALSE)
}
\arguments{
\item{Y}{Matrix of DNA methylation beta values (CpGs x subjects).  Missing values \emph{are} supported.}

\item{X}{Design matrix (subjects x covariates).}

\item{K}{Latent variable dimension (d in Houseman et al., 2013, technical report)}

\item{smallOutput}{Smaller output?  (Should be FALSE if you intend to run bootstraps.)}
}
\value{
A list object of class \dQuote{RefFreeEwasModel}. The most important elements are Beta and Bstar.
}
\description{
Reference-free method for conducting EWAS while deconvoluting DNA methylation arising as mixtures of cell types.
}
\details{
Reference-free method for conducting EWAS while deconvoluting DNA methylation arising as mixtures of cell types.  This method is similar to surrogate variable analysis (SVA and ISVA), except that it makes additional use of a biological mixture assumption.  Returns mixture-adjusted Beta and unadjusted Bstar, as well as estimates of various latent quantities.
}
\examples{
data(RefFreeEWAS)

if (interactive()) {
  tmpDesign <- cbind(1, rfEwasExampleCovariate)
  tmpBstar <- (rfEwasExampleBetaValues \%*\% tmpDesign \%*\% solve(t(tmpDesign)\%*\%tmpDesign))
  
  EstDimRMT(rfEwasExampleBetaValues-tmpBstar \%*\% t(tmpDesign))$dim  
}
test <- RefFreeEwasModel(
  rfEwasExampleBetaValues, 
  cbind(1,rfEwasExampleCovariate),
  4
)

testBoot <- BootRefFreeEwasModel(test,10)
summary(testBoot)
}
\seealso{
\code{BootRefFreeEwasModel}
}
\keyword{deconvolution,DNA}
\keyword{methylation,EWAS,surrogate}
\keyword{mixture,svd}
\keyword{variable,cell}
