% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEwasExampleBetaValues-dataset.R
\docType{data}
\name{rfEwasExampleBetaValues}
\alias{rfEwasExampleBetaValues}
\title{RefFreeEWAS Datasets: rfEwasExampleBetaValues}
\format{
1000 CpG sites x 250 subjects
}
\usage{
rfEwasExampleBetaValues
}
\description{
1000 CpG sites x 250 subjects.  First 250 CpGs are DMRs for the cell types, although the idea is that this would not be known in practice.
}
\keyword{datasets}
