% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEwasExampleCovariate-dataset.R
\docType{data}
\name{rfEwasExampleCovariate}
\alias{rfEwasExampleCovariate}
\title{RefFreeEWAS Datasets: rfEwasExampleCovariate}
\format{
Numeric vector of length 250
}
\usage{
rfEwasExampleCovariate
}
\description{
Simulated covariate for mixed-cell DNA methylation data set
}
\keyword{datasets}
