% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEwasExampleTRUEAlpha-dataset.R
\docType{data}
\name{rfEwasExampleTRUEAlpha}
\alias{rfEwasExampleTRUEAlpha}
\title{RefFreeEWAS Datasets: rfEwasExampleTRUEAlpha}
\format{
1000 intercept values
}
\usage{
rfEwasExampleTRUEAlpha
}
\description{
True alpha intercepts used in simulation; these may not match exactly due to cell mixtures.
}
\keyword{datasets}
