% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEwasExampleTRUEBeta-dataset.R
\docType{data}
\name{rfEwasExampleTRUEBeta}
\alias{rfEwasExampleTRUEBeta}
\title{RefFreeEWAS Datasets: rfEwasExampleTRUEBeta}
\format{
1000 coefficient values
}
\usage{
rfEwasExampleTRUEBeta
}
\description{
True beta coefficients used in simulation (for comparison purposes).
}
\keyword{datasets}
