% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEwasExampleTRUEMethDMR-dataset.R
\docType{data}
\name{rfEwasExampleTRUEMethDMR}
\alias{rfEwasExampleTRUEMethDMR}
\title{RefFreeEWAS Datasets: rfEwasExampleTRUEMethDMR}
\format{
1000 x 4 matrix
}
\usage{
rfEwasExampleTRUEMethDMR
}
\description{
True M matrix (cell-specific methylation values) used in simulation (for comparison purposes)
}
\keyword{datasets}
