% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEwasExampleTRUEOmega-dataset.R
\docType{data}
\name{rfEwasExampleTRUEOmega}
\alias{rfEwasExampleTRUEOmega}
\title{RefFreeEWAS Datasets: rfEwasExampleTRUEOmega}
\format{
250 x 4 matrix
}
\usage{
rfEwasExampleTRUEOmega
}
\description{
True Omega (cell mixture) coefficients used in simulation (for comparison purposes)
}
\keyword{datasets}
