\name{summary.BootRefFreeEwasModel}
\alias{summary.BootRefFreeEwasModel}
\title{summary.BootRefFreeEwasModel}
\description{
Summary method for objects of type BootRefFreeEwasModel; calculates bootstrap mean and standard deviation.
}
\usage{
\method{summary}{BootRefFreeEwasModel}(object,...)
}
\arguments{
  \item{object}{BootRefFreeEwasModel object to summarize}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{RefFreeEwasModel}} for example.
}
