% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefFreeCellMix.R
\name{summary.RefFreeCellMix}
\alias{summary.RefFreeCellMix}
\title{summary.RefFreeCellMix}
\usage{
\method{summary}{RefFreeCellMix}(object, ...)
}
\arguments{
\item{object}{RefFreeCellMix object to summarize.}

\item{...}{Unused.}
}
\description{
Summary method for objects of type RefFreeCellMix.
}
\details{
See \code{RefFreeCellMix()} for example.
}
