% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefFreeCellMix.R
\name{SVDwithMissing}
\alias{SVDwithMissing}
\title{SVD with missing values}
\usage{
SVDwithMissing(Y)
}
\arguments{
\item{Y}{Matrix for which to compute SVD.}
}
\value{
singular value decomposition (as returned by svd function).
}
\description{
Compute singular value decomposition on a matrix with missing values,
using a naive/simple method for imputing missing values by row-mean.
}
\details{
Computes singular value decomposition on a matrix with
missing values, using a naive/simple method for imputing missing
values by row-mean.  Not recommended for matrices with very large
numbers of missing values.
}
