\name{X.HNSCC.caseStatusAge}
\alias{X.HNSCC.caseStatusAge}
\docType{data}
\alias{X.HNSCC.caseStatusAge}
\title{HNSCC Example - Covariates}
\description{Case status (0=control, 1=case) and age (as Z-score) for HNSCC data set}
\usage{X.HNSCC.caseStatusAge}
\format{Numeric matrix of dimension 182 x 3}

