\name{Y.HNSCC.averageBetas}
\alias{Y.HNSCC.averageBetas}
\docType{data}
\alias{Y.HNSCC.averageBetas}
\title{HNSCC Example - DNA Methylation Average Betas}
\description{Peripheral blood from 92 head and neck squamous cell carcinoma (HNSCC) patients and 92 controls. 
GEO Accession #GSE32393 with 2 outlier cases removed.}
\usage{Y.HNSCC.averageBetas}
\format{Numeric matrix of dimension 26486 by 182}

