% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapPairs.R
\name{bootstrapPairs}
\alias{bootstrapPairs}
\title{One Bootstrap Sample for Pairs}
\usage{
bootstrapPairs(obs, pairID)
}
\arguments{
\item{obs}{Observation ids (numeric vector).}

\item{pairID}{Pair IDs (one unique value per pair).}
}
\value{
A vector of IDs corresponding to bootstrapped pairs
}
\description{
Bootstrap generation procedure for sampling paired data (e.g. twin data)
}
\details{
Generates one bootstrapped set of ids corresponding to pairs for
the method for conducting EWAS while deconvoluting DNA methylation
arising as mixtures of cell types.
Typically not run by user.
}
\seealso{
\code{BootRefFreeEwasModel}, \code{PairsBootRefFreeEwasModel}
}
\keyword{deconvolution,DNA}
\keyword{methylation,EWAS,surrogate}
\keyword{mixture}
\keyword{variable,cell}
