% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDimIC.R
\name{EstDimIC}
\alias{EstDimIC}
\title{Dimension estimation by AIC and BIC}
\usage{
EstDimIC(Rmat, Krange = 0:25)
}
\arguments{
\item{Rmat}{Residual matrix for which to estimate latent dimension.}

\item{Krange}{Vector of integers representing candidate dimensions to consider.}
}
\value{
A list containing AIC and BIC for candidate dimensions, as well as the best dimension for each.
}
\description{
Method for estimating latent dimension by AIC and BIC.
}
\details{
Method for estimating latent dimension by AIC and BIC.
Inferior to the RMT method in the isva package, but it appears here because it's mentioned in our paper.
}
\examples{
data(RefFreeEWAS)

if (interactive()) {
  tmpDesign <- cbind(1, rfEwasExampleCovariate)
  tmpBstar <- rfEwasExampleBetaValues \%*\% tmpDesign \%*\% solve(t(tmpDesign)\%*\%tmpDesign)

  EstDimIC(rfEwasExampleBetaValues-tmpBstar \%*\% t(tmpDesign))
}
}
\seealso{
\code{EstDimRMT}
}
