\name{ImputeByMean}
\alias{ImputeByMean}
\title{Simple imputation method based on row-mean}
\description{Simple method for imputing missing values by row-mean}
\usage{ImputeByMean(Y)}
\arguments{
  \item{Y}{Matrix to impute.}
}
\value{Matrix with missing values replaced by imputed values}

