\name{print.BootRefFreeEwasModel}
\alias{print.BootRefFreeEwasModel}
\title{print.BootRefFreeEwasModel}
\description{
Print method for objects of type BootRefFreeEwasModel
}
\usage{
\method{print}{BootRefFreeEwasModel}(x,...)
}
\arguments{
  \item{x}{BootRefFreeEwasModel object to print}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{RefFreeEwasModel}} for example.
}
