\name{print.RefFreeCellMix}
\alias{print.RefFreeCellMix}
\title{print.RefFreeCellMix}
\description{
Print method for objects of type RefFreeCellMix
}
\usage{
\method{print}{RefFreeCellMix}(x,...)
}
\arguments{
  \item{x}{RefFreeCellMix object to print}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{RefFreeCellMix}} for example.
}

