\name{print.RefFreeEwasModel}
\alias{print.RefFreeEwasModel}
\title{print.RefFreeEwasModel}
\description{
Print method for objects of type RefFreeEwasModel
}
\usage{
\method{print}{RefFreeEwasModel}(x,...)
}
\arguments{
  \item{x}{RefFreeEwasModel object to print}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{RefFreeEwasModel}} for example.
}
