% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRefFreeEwasModel.R
\name{print.summaryBootRefFreeEwasModel}
\alias{print.summaryBootRefFreeEwasModel}
\title{print.summaryBootRefFreeEwasModel}
\usage{
\method{print}{summaryBootRefFreeEwasModel}(x, ...)
}
\arguments{
\item{x}{summaryBootRefFreeEwasModel object to print}

\item{...}{Unused..}
}
\description{
Print method for objects of type summaryBootRefFreeEwasModel
}
\details{
See \code{RefFreeEwasModel} for example.
}
