% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefFreeCellMix.R
\name{RefFreeCellMixArrayDevianceBoots}
\alias{RefFreeCellMixArrayDevianceBoots}
\title{RefFreeCellMixArrayDevianceBoots}
\usage{
RefFreeCellMixArrayDevianceBoots(
  rfArray,
  Y,
  R = 5,
  EPSILON = 1e-09,
  bootstrapIterations = 5
)
}
\arguments{
\item{rfArray}{list of RefFreeCellMix objects (e.g. from RefFreeCellMixArray)}

\item{Y}{Methylation matrix on which x was based}

\item{R}{Number of bootstrapped vectors to return}

\item{EPSILON}{Minimum value of variance (zero variances will be reset to this value)}

\item{bootstrapIterations}{Number of RefFreeCellMix iterations to use in bootstrap}
}
\description{
Matrix of bootstrapped deviances corresponding to an array of reference-free cell-mixture decompositions
}
\details{
Matrix (multiple vectors) of bootstrapped deviances corresponding to an array of reference-free cell-mixture decompositions,
used to determine optimal number of cell types. This function returns one bootstrapped vector.
The bootstrapped deviance is based on normal distribution applied to errors of Y after accounting for cell mixture effect, Mu   Omega^T.
See \code{RefFreeCellMixArray} for example.
}
