\name{rfEwasExampleBetaValues}
\alias{rfEwasExampleBetaValues}
\docType{data}
\alias{rfEwasExampleBetaValues}
\title{Simulated mixed-cell DNA methylation data set}
\description{1000 CpG sites x 250 subjects.  First 250 CpGs are DMRs for the cell types, although the idea is that this would not be known in practice.}
\usage{rfEwasExampleBetaValues}
\format{1000 CpG sites x 250 subjects.}
