\name{rfEwasExampleCovariate}
\alias{rfEwasExampleCovariate}
\docType{data}
\alias{rfEwasExampleCovariate}
\title{Simulated covariate for mixed-cell DNA methylation data set}
\description{Vector of covariates corresponding to 250 subjects.}
\usage{rfEwasExampleCovariate}
\format{Numeric vector of length 250}
