% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootRefFreeEwasModel.R
\name{summary.BootRefFreeEwasModel}
\alias{summary.BootRefFreeEwasModel}
\title{summary.BootRefFreeEwasModel}
\usage{
\method{summary}{BootRefFreeEwasModel}(object, ...)
}
\arguments{
\item{object}{BootRefFreeEwasModel object to summarize.}

\item{...}{Unused.}
}
\description{
Summary method for objects of type BootRefFreeEwasModel;
calculates bootstrap mean and standard deviation.
}
\details{
See \code{RefFreeEwasModel} for example.
}
