\name{summary.RefFreeCellMix}
\alias{summary.RefFreeCellMix}
\title{summary.RefFreeCellMix}
\description{
Summary method for objects of type RefFreeCellMix.
}
\usage{
\method{summary}{RefFreeCellMix}(object,...)
}
\arguments{
  \item{object}{RefFreeCellMix object to summarize}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{RefFreeCellMix}} for example.
}

