% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.HNSCC.caseStatusAge-dataset.R
\docType{data}
\name{X.HNSCC.caseStatusAge}
\alias{X.HNSCC.caseStatusAge}
\title{HNSCC Datasets: X.HNSCC.caseStatusAge}
\format{
Numeric matrix of dimension 182 x 3
}
\usage{
X.HNSCC.caseStatusAge
}
\description{
Case status (0=control, 1=case) and age (as Z-score) for HNSCC data set
}
\keyword{datasets}
