% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y.HNSCC.averageBetas-dataset.R
\docType{data}
\name{Y.HNSCC.averageBetas}
\alias{Y.HNSCC.averageBetas}
\title{HNSCC Datasets: Y.HNSCC.averageBetas}
\format{
Numeric matrix of dimension 26486 by 182
}
\usage{
Y.HNSCC.averageBetas
}
\description{
Peripheral blood from 92 head and neck squamous cell carcinoma (HNSCC) patients and 92 controls.
}
\keyword{datasets}
